SQL to pl conversion, Reload page after conversion, XLSX and XLS integer error resolved

+++ AFleSe-master/flese/src/fileConverters/XLSXToCSVConverter.java
@@ -53,7 +53,7 @@ public class XLSXToCSVConverter {
 						break;
 
 					case Cell.CELL_TYPE_NUMERIC:
-						data.append(cell.getNumericCellValue());
+						data.append(((int) cell.getNumericCellValue()));
 						break;
 
 					case Cell.CELL_TYPE_STRING:
