+++ AFleSe-master/flese/src/fileConverters/XLStoCSVConverter.java
@@ -55,7 +55,7 @@ public class XLStoCSVConverter {
 						break;
 
 					case Cell.CELL_TYPE_NUMERIC:
-						data.append(cell.getNumericCellValue());
+						data.append(((int) cell.getNumericCellValue()));
 						break;
 
 					case Cell.CELL_TYPE_STRING:
